+++ src/providers/index.ts
@@ -33,6 +33,8 @@ export { usePermissionCheck } from "./hooks/usePermissionCheck";
 export { useRestrictedLink } from "./hooks/useRestrictedLink";
 export type { RestrictedLink } from "./hooks/useRestrictedLink";
 export { useRouteAccess } from "./hooks/useRouteAccess";
+export { useRelayPagination } from "./hooks/useRelayPagination";
+export type { RelayPageVars, RelayPagination } from "./hooks/useRelayPagination";
 export { useRouteInfo } from "./RouteProvider/hooks";
 export type { RouteContextValue } from "./RouteProvider/hooks";
 export type { RouteProviderProps, RouteTemplateProps } from "./RouteProvider";